Fix bug in deploy scripts.
Deploy scripts were deploying the root level files instead of the
`_static` directory. This was due to the working directory being lost when
having scripts call other scripts.

The naive approach was used to fix this: set working directories
explicitly in each directory that requires a new working directory.

Approaches that retain working directories may be possible with NPM
scripts, and these could be used in the future. For now, there is no
need for that level of granularity.